+++ marline_index/src/index/store/mod.rs
@@ -16,14 +16,14 @@ pub trait Store<H, S: Sketch>: Send + Sync
 where
     H: Clone + Send + Sync,
 {
-    /// hash → Sketch 
+    /// hash → Sketch
     fn get_sketch(&self, hash: &H) -> Result<Option<S>, IndexError>;
     fn put_sketch(&self, hash: &H, sketch: &S) -> Result<(), IndexError>;
 
     /// superfeature → Vec<Hash>
-    fn get_inverted(&self, tier: Tier, sf: u64) -> Result<Vec<H>, IndexError>;
-    fn put_inverted(&self, tier: Tier, sf: u64, hash: &H) -> Result<(), IndexError>;
-    fn remove_inverted(&self, tier: Tier, sf: u64, hash: &H) -> Result<(), IndexError>;
+    fn get_inverted(&self, tier: Tier, sf: u32) -> Result<Vec<H>, IndexError>;
+    fn put_inverted(&self, tier: Tier, sf: u32, hash: &H) -> Result<(), IndexError>;
+    fn remove_inverted(&self, tier: Tier, sf: u32, hash: &H) -> Result<(), IndexError>;
 
     ///
     fn len_sketches(&self) -> Result<usize, IndexError>;
